PMS. Reserva - Servicios. Cierre de Produccion. Marcar Servicio como borrado en lugar de borrarlo fisicamente en caso de que haya entrado en produccion (base imp produccion != 0). No tenerlo en cuenta en los procesos si tiene el flag de borrado.

+++ aon-ui-pms/src/main/java/com/esferalia/aon/ui/pms/controller/ProjectReservationServiceController.java
@@ -340,10 +340,18 @@ public class ProjectReservationServiceController extends LinesController {
 	public void onRemoveReservationService(ActionEvent event) throws ManagerBeanException {
 		ProjectReservationService reservationService = (ProjectReservationService)getTo();
 
+		if (reservationService.getProjectReservation().isDirty()) {
+			String message = "La Reserva ha sido modificada por otro usuario. Refrescar para obtener los datos actualizados.";
+			throw new AbortProcessingException(message);
+		}
+
 		ReservationUtils reservationUtils = new ReservationUtils();
     	reservationUtils.removeProjectReservationServiceDetails(reservationService);
-
-    	onRemove(event);
+    	if (!reservationService.isRemoved()) {
+    		onRemove(event);
+    	} else {
+    		onSearch(event);
+    	}
 
     	refreshReservationTotals();
 	}
